Removed the logic from the PasscodeViewController and added it to the manager
